Renamed provider folder
