+++ apexPropertiesProvider/provider/helper/ExtensionHelper.js
