+++ apexPropertiesProvider/provider/helper/OpenDialogLabel.js
