+++ apexPropertiesProvider/provider/helper/validateXML.js
