+++ apexPropertiesProvider/provider/parts/events/EventTypeProps.js
