+++ apexPropertiesProvider/provider/parts/events/TerminateEventProps.js
