+++ apexPropertiesProvider/provider/parts/message/SimpleMessageProps.js
