+++ apexPropertiesProvider/provider/parts/pageItems/PageItemProps.js
