+++ apexPropertiesProvider/provider/parts/processVariables/ProcVarGroup.js
