+++ apexPropertiesProvider/provider/parts/processVariables/ProcVarProps.js
