+++ apexPropertiesProvider/provider/parts/timer/CustomTimerProps.js
