+++ apexPropertiesProvider/provider/plugins/monacoEditor.js
