+++ index.js
@@ -4,7 +4,7 @@ import colorPickerModule from 'bpmn-js-color-picker';
 import BpmnModeler from 'bpmn-js/lib/Modeler';
 import bpmnlintConfig from './.bpmnlintrc';
 import apexModdleDescriptor from './apexPropertiesProvider/descriptor/apexProps';
-import propertiesProviderModule from './apexPropertiesProvider/new_provider';
+import propertiesProviderModule from './apexPropertiesProvider/provider';
 import customPaletteProviderModule from './custom/palette';
 import bpmnDiOrdering from './modules/bpmnDiOrdering';
 import drilldownCentering from './modules/drilldownCentering';
